+++ lib/src/autolink_utils.dart
@@ -1,10 +1,10 @@
-import 'package:selectable_autolink_text/selectable_autolink_text.dart';
+import 'link_attr.dart';
 
 class AutoLinkUtils {
   AutoLinkUtils._();
 
   static const urlRegExpPattern =
-      r'https?://([\w-]+\.)+[\w-]+(/[\w-./?%&=~#+]*)?';
+      r'https?://([\w-]+\.)+[\w-]+(/[\w-./?%&@\$=~#+]*)?';
   static const phoneNumberRegExpPattern = r'[+0]\d+[\d-]+\d';
   static const emailRegExpPattern = r'[^@\s]+@([^@\s]+\.)+[^@\W]+';
   static const defaultLinkRegExpPattern =